rag-battle-3/li/requirements.txt
llama-index
openai
pinecone-client
llama-index-vector-stores-pinecone